refactor: relocate manifest (API)
