+++ src/__mocks__/manifest.ts
@@ -1,5 +1,6 @@
+import type { Manifest } from "../api/manifest";
 import { Version } from "../common/version";
-import type { getManifest as __getManifest, getSoftwareMinimumVersion as __getSoftwareMinimumVersion, Manifest } from "../manifest";
+import type { getManifest as __getManifest, getSoftwareMinimumVersion as __getSoftwareMinimumVersion } from "../manifest";
 
 /**
  * Mock {@link Manifest}.
