+++ src/__tests__/i18n.test.ts
@@ -2,9 +2,9 @@ import fs, { Dirent } from "node:fs";
 import path from "node:path";
 
 import { getMockedLogger } from "../../tests/__mocks__/logging";
+import type { Manifest } from "../api/manifest";
 import { I18nProvider } from "../i18n";
 import type { Logger } from "../logging";
-import type { Manifest } from "../manifest";
 
 describe("I18nProvider", () => {
 	/**
